Runtime: Create basic runtime objects
This commit includes the Agent, Realm, ExecutionContext, and EnvRecord
objects. These are the fundamental objects in the ECMA spec, and will
power the runtime of this JS engine.

The spec of course explicitly mentions that these concepts don't have
to have a direct equivalent in an implementation, but following the
spec as rigidly as possible will make it easier to get the project off
the ground. Once it can actually _run_ code, alternate systems can be
considered.

This commit won't compile, as this and the next one were written
together, however I still wanted to separate them for clarity.