Parser/AST: Completely rewrite the parser
I didn't like ripping this from Serenity's LibJS, so I decided to write
my own. This parser, while more inefficient, is designed to strictly
follow the spec. Anyone reading the parser can find the corresponding
section in the spec with ease. It is obviously not done, but the most
important parts are.

Additionally, this commit comes with a Static Semantic system. They'll
be implemented via interfaces on the individual AST nodes. These will
make compiling the AST tree much easier.