AST: Reworks Static Semantics
The previous interface method wasn't really realizable for two reasons:

1. Every chain production (nonterminal with only one other nonterminal
on its right-hand side) automatically received an implicit
implementation of every SS with delegates it to its children
nonterminal.
2. All nodes have a default implementation of Contains

So now, because of 1., all ASTNodes implements all SS functions
automatically, and can be overridden with actual implementations if
necessary. The ASTNode also receives in its constructor all of its
children for the implementation of 2.