AST: Rewrite node structure
...again. This isn't the first time, and it probably won't be the last.

This patch gets rid of all of the useless wrapper nodes, like Literal,
and makes all of the fundamental node types (StatementNode,
ExpressionNode, etc.) into interfaces. ASTNode is also now an interface,
with an impl called NodeBase.

Unfortunately this change messed up a lot of SS stuff, so fixed for that
will be included in the following commits.