Runtime: More systematic changes
This further reworks the agent/execution context/env/realm system, which
from this point forward I will refer to as the AEER. It's hard to
describe exactly what I did in this commit, but I did it to rework the
public API in a way that makes more sense and is more inline with the
spec. Look at JSThing.kt for examples (I really need to come up with a
project name...).