Compiler/Runtime: Add error support!
This is a massive change. I wish I could split it into multiple commits,
but that's not really possible.

All native methods that can throw exceptions are now annotated with
@JSThrows. When the compiler calls the "operation" extension method, it
automatically searched for the method in Operations.kt, finds the
method, and if it is annotated with @JSThrows, it inserts an exception
check.

Exceptions are stored in the execution context object. The way they work
with regards to the AEER stack is still a bit up in the air, and can
change in the future, but this is a solid start. I will learn more as I
start to implement 'throws' and 'try-catch'