Compiler: Add "Extra Methods" concept; fix function bug
An "extra method" is a method created on a compiled class instance for
the purpose of "capturing" return statements. For example, inside of a
try-statement, any returns that occur due to an exception (i.e. from a
call to Agent.hasError()) need to not halt executing of the
try-statement's surrounding context.

To solve this, the try-statement puts its try-block inside of a separate
member function within the class. This way, it can call the method, make
its own call to Agent.hasError(), and act occordingly.

This same idea will also be using in various other structures, lexical
for-statements being the most prominant.

This also fixes a bug with functions where its envRecord property was
not being set. This resulting in any code inside of a function not
having access to the global object.