Add an interpreter
This does the same thing as the Compiler, but in a different way. This
will be the primary way that Reeva is executed going forward, at least
for the time being. I'm not sure if Compiling to bytecode actually
results in any meaningfull performance boost. When the interpreter gets
far along to actually run test262 tests, we will have to get the
compiler up to speed and check if the performance boost is worth it.