Parser: Improve parsing of CPEAAPL
I didn't really understand what the CPEAPPL node was or how it
functioned within the grammar. However, the V8's ECMAScript spec blog
part 4 is focused on this topic
(https://v8.dev/blog/understanding-ecmascript-part-4).

The CPEAAPL node shouldn't _actually_ be a node, and it should appear in
the compiler or interpreter. It is merely a parser construct, and should
be handled at the site where it is needed.