Core: Rewrite agent/realm system
This is a large commit that completely refactors how Agent work, as well
as the execution loop. Now, every thread has an associated Agent, and
Agents no longer have to be explicitly created. They can be accessed
statically (as they are stored in ThreadLocal storage), and do not have
to be propogated by the other objects (realms, contexts, etc).

Also, the Agent now has an event loop (kinda). This is to prepare for
the future with promises, setTimeout, etc. It should be easier to manage
these tasks and display asynchronous behavior in a single-threaded
engine.