All: Use java exceptions instead of tracking them manually
For some reason, I got the idea in my head that exceptions would be slow
for something like this, but I really disliked the previous style of
checking after almost every method call, so I transitioned to regular
exceptions to compare performance. Turns out, this is actually faster by
around 10-20%, and it looks way nicer.