Lexer/Parser/Interpreter: Remove private fields
After having read through a bunch of the private field spec issues, it
seems like the spec is still in a controversial state. Unfortunately,
the spec also covers fields in general, not just private fields. So I'm
basically implementing half of the spec, and ignoring the other half...

Also, the spec is very confusing in general. I'll probably be
implementing the rest of public fields on my own, mostly ignoring the
spec.