All: Add ES6 modules!
This is a massive change, and doesn't quite work perfectly, but I want
to commit this since it is a nice "checkpoint".

First and foremost, this commit adds parsing of import and export
declarations. Secondly, I've added a module system that is as close to
the spec as possible. Named exports, as well as aliases for those
exports, work great. Default exports don't work, as there is an issue
somewhere with creating the binding before initializing it. export from
statements don't work at all.