Runtime: Add Date objects
This was surprisingly easy to implement with Java's datetime library --
even string parsing, which I though was going to be a nightmare. Turns
out, DateTimeFormatterBuilder is absolutely amazing and does everything
I want it to do.

This implementation passed 83% of the tests in built-ins/Date. It could
easily pass more, but I'd like to work on something else for now