Runtime: Implement better JSValue -> JVM type conversion
This patch implements much better support for js -> jvm argument
conversion. It implements a heuristic algorithm to match the best
fitting jvm method for any given argument set, based loosely off
of the LiveConnect3 specification. Additionally, a new
test suite has been added to test the functionality. Currently
it only has unit tests for type coercion, and end-to-end tests
are still required.

There is still more work to be done in this area, primarily
updating the JVM -> JSValue algorithms, as well as converting
from more complex types, such as Map or Set, but this is a great start.