Module: Add JVM classes to import statement; generalize ModuleRecord
This commit first and foremost generalized the ModuleRecord. Previously,
it had a hard dependency on a ModuleEnvRecord, and would do all of its
resolution through that env record. That dependency has been moved to
CyclicModuleRecord, which makes more sense as a CyclicModuleRecord
represents a module derived from actual JS text.

This change allows the seemless addition of the JVMPackageModuleRecord,
which allows for imports such as the following:

    import { ClassName } from 'jvm:com.example.examplepackage';

Stay tuned for namespace importing :)