Runtime: Add global jvm() function for extending JVM classes/interfaces
This is initial support for the following syntax:

    class Foo extends jvm(MyBaseClass, MyInterface1, MyInterface2) {
        // ...
    }

This will be an extremely nice alternative to Rhino's and Nashorn's
system of Java.extends(), which itself is extremely limited. This syntax
will eventually be almost identical to extending the JVM class in a JVM
language, but for now it is very limited.