Runtime: Add object shapes
Every object now has an associated Shape, which stores property keys in
the form of a linked-list-like hierarchy, as well as an offset of that
property into the object's property array (not map). This reduces key
duplication across objects with similar shapes.

This absolutely destroys performance, but hopefully we can improve that
in future commits.