Runtime: Refactor JSFunction to only have one abstract method
Now all functions simply have an evaluate function which takes a list of
values, which is called for both normal JS calls and construction calls.
'thisValue' and 'newTarget' are now functions on JSFunction, which are
automatically bound by the respective call or construct calls.

Previously, the spec's requirements for calling and constructing were
only being done on JSRuntimeFunctions, but they are now done for all
functions.