Runtime: Use strings for annotation attributes
This is a lot more compact than the bit fields. Now, each attribute
"part" is assigned a letter. A lowercase letter means the attribute is
specified as false (i.e. HAS_XYZ, but not XYZ), and an uppercase letter
means the attribute is specified as true (i.e. HAS_XYZ and XYZ). The
letters are:

    'C' - configurable
    'E' - enumerable
    'W' - writable

Some attribute examples:

    "ceW" -> { configurable: false, enumerable: false, writable: true }
    "E" -> { enumerable: true }
    <empty> -> { }