JSObject: Support full array indices range; use Longs for array indices
This commit supports the full extended range of array lengths.
Previously, the maximum array index allowed in an object's indexed
storage was 2^31-1, but it is not 2^32-1. We also now use Longs in many
places regarding arrays instead of Ints.

Unfortunately, as this is a large change, it introduces quite a few
regressions. These will hopefully be fixed in future commits