Test262Runner: Track test results manually and output to JSON
IntelliJ isn't able to export results on my machine, as they are too
large, and even if it could (it used to when I ignored most of the test
suite), all tests have timestamps. These timestamps mean that I can't
just throw them in a text-diff tool and compare results, as every test
will be different. It's easier to just track them myself.