AST/ScopeAnalyzer: Track whether variable declarations are inlinable
At the moment, the only reason a variable wouldn't be inlinable is if
there is a function in or below the current scope which has a reference
to the declaration. There may be other cases where we can't inline
variables, such as if the scope contains eval calls, but those cases can
be handled later.