Remove the interpreter and compiler
The new IR stage will require a completely new interpreter and compiler,
written from the ground up. As such, the current interpreter and
compiler have been removed in favor of their future, hopefully faster
versions.

The interpreter is unfortunately very deeply integrated into
the project, directly by name, so this commit includes a lot of randomly
commented out code. A better interface for the interpreter and compiler
should be created which is backend-dependent