AST: Completely overhaul structure
When I started this project, strict spec compliance was a primary goal.
However, after observing the speed of the engine... I have changed my
mind. Strict spec compliance is no longer a goal, and as such, the AST
tree can be drastically simplified. Many of the ASTNode functions have
been removed, and will be replaced as they are needed.