IR: Begin outputting sensible IR code
This is the first commit where the IRTransformer actually emits opcodes!
This commit is a bit messy, as it took quite a lot to achieve this.
Mostly changes in Variable/VariableSourceNode semantics, most notable
that VariableSourceNode is now only a single source instead of multiple
sources. I initially made it multi-source for destructuring, but I was
getting a bit ahead of myself. I can look into it again when I get that
far.