IR: Being adding support for non-inlineable variables
This doesn't quite work due to a quirk of EnvRecords (the first
time a value is assigned to a variable, one has to explicitly call
intiializeBinding, which is tough to keep track of). In order to make
it work properly, EnvRecords will have to be refactored. This will be a
lot of work, and needs its own commit :)