Parser/Lexer/AST: Heavily refactor
The parser/Lexer is now 2x as fast, and is also much nicer. It comes
with a heavily overhauled AST to reduce clutter. Additionally, the
parser is now responsible for initializing scopes, which remove the
extra "ScopeAnalyzer" step (its faster too, since we don't have to
re-traverse the AST tree).