Parser: Fix node ranges
This isn't perfect, but it's definitely solid. We now select the
correct ending position in the "nps" function. Additionally, most of
the position bugs were due to accidental non-local returns inside of
an nps expression. The lambda in nps has now been made crossinline
to prevent that from ever happening again.