Core: Restructure the public API a bit
This is a step forward in finalizing the relationship between all of
the fundamental objects in Reeva. Here's the plan now:

At the start and end of a program, Reeva.{setup,teardown}() must be
called respectively. This is no different than before.

The Reeva class owns (i.e. stores) all of the created Agent objects.
There is at any time one single active Agent in any thread. Multiple
threads can have their own Agent. Inter-Agent communication is a
problem for another day.

Every agent has its own execution stack. The ExecutionContext on the
top of the stack is the currently executing execution.
ExecutionContext objects can be pushed onto this stack. Every time a
function is called (any JSFunction, not just script functions), a new
ExecutionContext is pushed onto the stack.

The ExecutionContext object is pretty useless right now, so I might
try and figure out a way to remove it in the future. But for now, it
stores the current Realm, which is heavily used by the runtime.

The IRInterpreter is a stateless class, and can be called at any time.
Also, it has been refactored to operate on IRFunctions only. To
support this, top-level scripts now compiled into an IRFunction.

This is still pretty volatile, and will probably be modified further
in the future.