IRTransformer: Transform try-finally statements
This commit does not handle try-catch or try-catch-finally statements.
The generated IR for try-finally statements more closely resembles the
JVM rather than V8, as we will eventually compile JS try-catch-finally
statements to JVM try-catch-finally statements, so if they are similar
it makes my life easier in the future.

Currently, the handler ranges that are generated are a bit different
than the ranges generated by the JVM. This may have to change in the
future.