IRTransformer: Improve handling of function parameters
Previously, when a function was defined, the parameter were being
assigned to env slots before a new env had actually been pushed (as it
was relying on a call to visitBlock to push the env, after the
parameters had been handled).

Now, we have a nice helper to enter and exit scopes, and we explicitly
call super.visitBlock inside of the function helper method to avoid the
extra scope push that visitBlock would do.

This commit also improves the numbering of env variables, as non-
inlineable parameters would mess up the ordering.