IRTransformer: Begin adding {function,global}DeclarationInstantiation
The transformer _sort of_ had this previously, but for some reason it
was only for functions, so it didn't affect the top-level script. Now,
var declarations are properly hoisted without their values and
functions are properly hoisted _with_ their values.

This also adds some opcodes which will be handled in future commits.