Scope: Do not link variable until after parsing
Consider the following code:

	var f = 10;

	function foo() {
    	console.log(f);
    	var f = 20;
    }

When we are parsing the console.log statement, it is not possible to
know that that variable is declared in the current scope, so the
reference is linked to the global 'f' declaration. However, it should
be linked to the local variable declaration.

This is trivial _after_ parsing, and linking variables during parsing
doesn't really matter since all unlinked variables are processed at the
end anyways. So, scopes no longer link variables as they parse; rather,
all variables are linked at the end.