Drastically simplify the opcode
I'm not sure why I made the opcode class so strange, but this is a
much more sensible way to do it. In the theme of removing stuff, this
commit removes the FeedbackVector and PeepholeOptimizer. It also
introduces Blocks, which are managed by the Generator. I've learned
from LibJS that doing this upfront is much easier than waiting until
optimization, and it also makes the code look *so* much better.