Transformer/Interpreter: Support try-catch-finally statements
Each block can have an associated handler, which the interpreter will
jump to if an exception occurs inside of that block. Finally blocks
are implemented with a jump table, where a secondary register is used
to tell the finally block if it should fallthrough to another block,
throw the value in a scratch register, or return the value in a scratch
register.