Realm/Interpreter: Revert to a linked-list-based EnvRecord model
The stack-based approach is very nice in that we can get an offset
EnvRecord for free, but it has one fatal flaw which I overlooked when
I initially designed this: Captures! The bane of all JS VM devs
everywhere.

The Realm has to store the current lexical and variable EnvRecords
during execution, but each individual EnvRecord needs to store its
outer record at _creation time_, so that it can reference captured
variables.