Agent: Process microtasks at the end of Agent::run
This will process Promises which are called as a direct result of
running a script. I'm not sure how to process microtasks more
generally, i.e. after Agent::run.

For example, suppose you have the following script:

    function foo() {
    	return new Promise(resolve => resolve(10));
    }

If you run this script, and then on the JVM side grab the foo function
and call it, the result will never resolve because the microtasks will
never be processed. I can think of two solutions:

  - Force the users of Reeva to explicitly churn the microtask queue
    if they need to. This has the advantage of being much simpler for
    me, but of course has the disadvantage of being more complex and
    less intuitive for the user.
  - Introduce more microtask checkpoints in the codebase. I'm not sure
    where these would go, perhaps at the end of JSFunction::evaluate
    behind a processMicrotasks: Boolean flag or something.