GlobalScope: Remove the requiresEnv() = false override
The global scope can not contain slots, as it is created by the Realm
well before even the AST is parsed. If the global scope needs slots,
that much be achieved by pushing a declarative env record. Note that
this doesn't strictly follow the spec; the spec says the global env
record consists of an object env _and_ a decl env.