Scope/Transformer: Initial support for arguments objects
The most important part of this commit is moving the "does this
function need an arguments object" logic from the transformer to the
ScopeResolver. This is necessary because the transformer needs to be
able to treat references to the arguments object like any other
declared variable (meaning it can be inlined as well).

Mapped arguments aren't fully supported yet because there is no way to
lookup a name's id in an env record. This will need some additional
work