Scope: Treat receiver as a VariableSourceNode; Add new.target support
The receiver of a function can be captured in closures (i.e. arrow
functions, which don't introduce their own receiver binding). Having
the receiver as a source node allows this to be possible.

And since new.target is similar to the receiver, I added support for it
in this commit as well. It does not need to be a variable because it
cannot be captured (new.target in an arrow function is a syntax error).