Agent: Use a background Thread for microtasks
This isn't an event queue (I think...), it just allows Reeva to
continuously run microtasks in the background as we manually
evaluate macrotasks. I'm quite unsure if this is still spec-compliant,
or how we might turn this into a real event loop with macrotasks, but
that is for future me to figure out.