Transformer/Interpreter: Support class fields
Class fields are implemented much like they are in V8. If a class has
instance fields, a function is created which installs them on the
receiver of that function. This function is called properly in all
of the different class constructor configurations:

  - Base class, explicit constructor
  - Base class, implicit constructor
  - Derived class, implicit constructor
  - Derived class, explicit constructor

In the derived classes, the field function is called directly after
the super() call. In base classes, it is called immediately. In order
to support these changes, the LdaClosure opcode had to be added. This
loads the accumulator with the object which has the field function
defined on it (via an internal symbol).

Static fields are initialized directly on the class constructor
function.