Transformer: Replace scuffed FindLoops algorithm
This replaces the old algorithm, which I just came up with on the spot,
with a proper algorithm based on a paper which uses a dominator tree.
The old algorithm was exponential time, and would blow up with large
scripts (even if they were completely linear with no loops!). It is now
much more complex, but also much more efficient.