Parser: Rework arrow functions, allow identifier param with no parens
This gets rid of the terrible CPEAAPLVisitor system I had before. That
was required because we had side effects at the time (the parser set
up the scopes), however now that we resolve scopes after parsing,
we have the ability to rewind the parser. This allows us to treat an
open paren as an arrow function, and then simply go back if it doesn't
work out.