Runtime: Make Builtin more generic
This makes Builtin a simple interface that provides a MethodHandle,
allowing library users to make their own Builtins. This also removes
the defineNativeFunction method from JSObject, meaning in order to
attach a native function, a Builtin must be used. This enforces the
use of MethodHandles, which improves speed and will later be used for
optimization.