Start another rewrite (again)
I've learned a lot from my adventures into IR/compiling/graphing, and
something I've taken away is that shoehorning something that works (V8)
into a completely different environment (stack-based JVM) just doesn't
work. Someone smarter than me could probably pull it off, but it's
gotten to the point where I don't want to work on Reeva anymore because
it's too complex. So, I'm going to try to rewrite the entire
architecture to be stack-based.

I also thought in the past that designing the system with optimization
in mind from the start was a good thing, but in reality it just made
things way too complicated. So, I'm not going to be thinking about
optimization at all. My goal right now is to parse the AST into IR,
and go straight from IR to bytecode.

...Ok, maybe I'm thinking a little about optimization. If I *really*
wasn't, I wouldn't bother with an IR stage at all. So just a little.