Interpreter: Rename lexicalEnv to activeEnvRecord
The name "lexicalEnv" is a remnant from the days where Reeva had both a
lexical environment (for let/const) and a var environment (for vars).
This is the way the spec specifies it, but it was cumbersome to deal
with, so it was ditched in favor of the parser-time hoistable scope
resolution mechanism.