Parser: Store processed and raw identifier string in IdentifierNode
Some places in the Parser require the raw name as it appeared in the
source file, and other places require the processed name (with escape
sequences collapsed). So now the IdentifierNode parsing function does
the string unescaping, and returns both in the IdentifierNode. This
fixes a bunch of identifier escaping-related parser errors.