All: Make the internal execution api a bit more ergonomic
The Executable class was pretty terrible; it tried to cover all use
cases for all situations, when in practice a sealed class/interface
or other target-typed structure would be much better. So that's what
this commit does! Each execution stage (parser, transformer,
interpreter) consumes and returns only what they are interested in.

This improves code quality by quite a bit.