ModuleRecord: Split into the three base classes
Turns out ECMAScript made a good distinction here. I want to add JVM
modules, and in order to do that, I'll need to make a new subclass of
ModuleRecord. It certainly won't be cyclic, so I needed the ModuleRecord
class to be fully generic, as it is in the spec. I didn't really need to
split SourceTextModuleRecord and CyclicModuleRecord, but I was already
refactoring, so why not.