JVMCompat: Fix JSClass{Instance,}Object prototype heirarchy
The prototype hierarchy for jvm class objects was completely wrong, so
this commit puts everything where it's supposed to be. Additionally, it
fixes the string method locations, and allows one to print both class
objects and class instances.

This commit also introduces JSClassInstanceObject.wrap() to turn an
arbitrary JVM object into a JSValue.