All: Track call stack location information
This commit adds a dedicated CallStack class, which keeps track of the
current call stack, along with source location information. In order to
do this, the transformer keeps the location information of opcodes of
interest (current calls, constructs, and throws). These gets relayed to
the interpreter, which sets a pending location in the CallStack class.