All: Tweak Agent/Realm model a bit
This commit rewrites the execution stack to look a bit more like the
execution model in the spec. Agent now has an execution stack populated
with ExecutionContext objects. I am also planning on relaxing the
thread model a bit, and part of this is allowing any piece of the code
to access execution state statically via the Agent. Because
ExecutionContext objects have Realm, this removes the need to pass
around Realms to every function (which is why this commit is so huge).

This comes with a lot of regressions, mostly to do with function call
semantics. These will hopefully be fixed in future commits!