Runtime: Remove realm as a field from JSObject
Objects don't really need a long-lasting reference to their realm.
So far, the only objects which do is JSFunction. Unfortunately we still
have to construct JSObject with a Realm as the constructor needs a
reference to the empty shape. This empty shape cannot be made static,
as all objects which use that shape would be kept alive by the static
reference. Perhaps this is something can be investigated in the future.