Runtime: Unify the two declarative env record classes
Turns out that splitting these was a terrible decision, since a
FunctionEnvRecord needs to be added, which is a Declarative Env Record
by spec terms, but in Reeva needs to support both optimized and
unoptimized modes. This commit also adds some logic for eval tainting,
which is of course the primary motivation for having an unoptimized
version of declarative env record.